+++ addons/flashlight/CfgWeapons.hpp
@@ -1,48 +1,26 @@
 class CfgWeapons {
+    class acc_flashlight;
+    class cup_acc_anpeq_15_flashlight_black_l;
+    class cup_acc_anpeq_15_flashlight_od_l;
+    class cup_acc_anpeq_15_flashlight_tan_l;
+    class cup_acc_anpeq_15_top_flashlight_black_l;
+    class cup_acc_anpeq_15_top_flashlight_od_l;
+    class cup_acc_anpeq_15_top_flashlight_tan_l;
+    class CUP_acc_CZ_M3X;
+    class cup_acc_flashlight;
+    class CUP_acc_Flashlight_desert;
+    class CUP_acc_Flashlight_wdl;
+    class CUP_acc_LLM01_L;
+    class cup_acc_llm_black;
+    class CUP_acc_MLPLS_Laser;
     class InventoryFlashLightItem_Base_F;
 
-    class acc_flashlight;
-    class CLASS(UTG_Flashlight): acc_flashlight {
-        displayName = "UTG Defender 126 (GT Edition)";
-        MACRO_MULTI_1
-
-        class ItemInfo: InventoryFlashLightItem_Base_F {
-            mass = 4;
-            class Flashlight {
-                MACRO_HI_FLASHLIGHT_SETTINGS
-            };
-        };
-        MRT_SwitchItemNextClass = QCLASS(UTG_Low_Flashlight);
-        MRT_SwitchItemPrevClass = QCLASS(UTG_Low_Flashlight);
-    };
-
-    class CLASS(UTG_Low_Flashlight): acc_flashlight {
-        displayName = "UTG Defender 126 (GT Edition)";
-        MACRO_MULTI_2
-
-        class ItemInfo: InventoryFlashLightItem_Base_F {
-            mass = 4;
-            class Flashlight {
-                MACRO_LOW_FLASHLIGHT_SETTINGS
-            };
-        };
-        MRT_SwitchItemNextClass = QCLASS(UTG_Flashlight);
-        MRT_SwitchItemPrevClass = QCLASS(UTG_Flashlight);
-    };
-
-    #include "Surefire_Flashlights.hpp"
-
-    // Multi-Function Laser/Lights - Top
-    #include "ANPEQ_Top_Black.hpp"
-    #include "ANPEQ_Top_Tan.hpp"
-    #include "ANPEQ_Top_OD.hpp"
-
-    // Multi-Function Laser/Lights - Side
-    #include "ANPEQ_Black.hpp"
-    #include "ANPEQ_Tan.hpp"
-    #include "ANPEQ_OD.hpp"
-    #include "LLM.hpp"
-
-    // Pistol based Laser/Lights
-    #include "Handgun.hpp"
+    #include "AN_PEQ.hpp"
+    #include "AN_PEQ_Top.hpp"
+    #include "CZ_M3X.hpp"
+    #include "LLM01.hpp"
+    #include "LLM_MKIII.hpp"
+    #include "MLPLS.hpp"
+    #include "Surefire.hpp"
+    #include "UTG.hpp"
 };
