+++ addons/flashlight/config.cpp
@@ -6,29 +6,32 @@ class CfgPatches {
         units[] = {};
         weapons[] = {
             QCLASS(ANPEQ_15_Laser_Black),
-            QCLASS(ANPEQ_15_Light_Black),
-            QCLASS(ANPEQ_15_Low_Light_Black),
-            QCLASS(ANPEQ_15_Vis_Laser_Black),
             QCLASS(ANPEQ_15_Laser_OD),
-            QCLASS(ANPEQ_15_Light_OD),
-            QCLASS(ANPEQ_15_Low_Light_OD),
-            QCLASS(ANPEQ_15_Vis_Laser_OD),
             QCLASS(ANPEQ_15_Laser_Tan),
+            QCLASS(ANPEQ_15_Light_Black),
+            QCLASS(ANPEQ_15_Light_OD),
             QCLASS(ANPEQ_15_Light_Tan),
+            QCLASS(ANPEQ_15_Low_Light_Black),
+            QCLASS(ANPEQ_15_Low_Light_OD),
             QCLASS(ANPEQ_15_Low_Light_Tan),
-            QCLASS(ANPEQ_15_Vis_Laser_Tan),
             QCLASS(ANPEQ_15_Top_Laser_Black),
-            QCLASS(ANPEQ_15_Top_Light_Black),
-            QCLASS(ANPEQ_15_Top_Low_Light_Black),
-            QCLASS(ANPEQ_15_Top_Vis_Laser_Black),
-            QCLASS(ANPEQ_15_Top_Laser_Tan),
-            QCLASS(ANPEQ_15_Top_Light_Tan),
-            QCLASS(ANPEQ_15_Top_Low_Light_Tan),
-            QCLASS(ANPEQ_15_Top_Vis_Laser_Tan),
             QCLASS(ANPEQ_15_Top_Laser_OD),
+            QCLASS(ANPEQ_15_Top_Laser_Tan),
+            QCLASS(ANPEQ_15_Top_Light_Black),
             QCLASS(ANPEQ_15_Top_Light_OD),
+            QCLASS(ANPEQ_15_Top_Light_Tan),
+            QCLASS(ANPEQ_15_Top_Low_Light_Black),
             QCLASS(ANPEQ_15_Top_Low_Light_OD),
+            QCLASS(ANPEQ_15_Top_Low_Light_Tan),
+            QCLASS(ANPEQ_15_Top_Vis_Laser_Black),
             QCLASS(ANPEQ_15_Top_Vis_Laser_OD),
+            QCLASS(ANPEQ_15_Top_Vis_Laser_Tan),
+            QCLASS(ANPEQ_15_Vis_Laser_Black),
+            QCLASS(ANPEQ_15_Vis_Laser_OD),
+            QCLASS(ANPEQ_15_Vis_Laser_Tan),
+            QCLASS(CZ_M3X_Laser),
+            QCLASS(CZ_M3X_Light),
+            QCLASS(CZ_M3X_Vis_Laser),
             QCLASS(LLM01_Laser_Black),
             QCLASS(LLM01_Light_Black),
             QCLASS(LLM01_Low_Light_Black),
@@ -37,19 +40,16 @@ class CfgPatches {
             QCLASS(LLM_Light_Black),
             QCLASS(LLM_Low_Light_Black),
             QCLASS(LLM_Vis_Laser_Black),
+            QCLASS(MLPLS_Laser),
+            QCLASS(MLPLS_Vis_Laser),
             QCLASS(Surefire_Flashlight_Black),
-            QCLASS(Surefire_Low_Flashlight_Black),
             QCLASS(Surefire_Flashlight_OD),
-            QCLASS(Surefire_Low_Flashlight_OD),
             QCLASS(Surefire_Flashlight_Tan),
+            QCLASS(Surefire_Low_Flashlight_Black),
+            QCLASS(Surefire_Low_Flashlight_OD),
             QCLASS(Surefire_Low_Flashlight_Tan),
             QCLASS(UTG_Flashlight),
-            QCLASS(UTG_Low_Flashlight),
-            QCLASS(CZ_M3X_Laser),
-            QCLASS(CZ_M3X_Light),
-            QCLASS(CZ_M3X_Vis_Laser),
-            QCLASS(MLPLS_Laser),
-            QCLASS(MLPLS_Vis_Laser)
+            QCLASS(UTG_Low_Flashlight)
         };
         magazines[] = {};
         requiredVersion = REQUIRED_VERSION;
@@ -61,61 +61,6 @@ class CfgPatches {
     };
 };
 
-class asdg_SlotInfo;
-class asdg_FrontSideRail: asdg_SlotInfo {
-    class compatibleItems {
-        tacgt_Surefire_Low_Flashlight_Tan = 1;
-        tacgt_ANPEQ_15_Light_Black = 1;
-        tacgt_ANPEQ_15_Low_Light_Black = 1;
-        tacgt_ANPEQ_15_Laser_Black = 1;
-        tacgt_ANPEQ_15_Vis_Laser_Black = 1;
-        tacgt_ANPEQ_15_Light_OD = 1;
-        tacgt_ANPEQ_15_Low_Light_OD = 1;
-        tacgt_ANPEQ_15_Laser_OD = 1;
-        tacgt_ANPEQ_15_Vis_Laser_OD = 1;
-        tacgt_ANPEQ_15_Light_Tan = 1;
-        tacgt_ANPEQ_15_Low_Light_Tan = 1;
-        tacgt_ANPEQ_15_Laser_Tan = 1;
-        tacgt_ANPEQ_15_Vis_Laser_Tan = 1;
-        tacgt_ANPEQ_15_Top_Laser_Black = 1;
-        tacgt_ANPEQ_15_Top_Light_Black = 1;
-        tacgt_ANPEQ_15_Top_Low_Light_Black = 1;
-        tacgt_ANPEQ_15_Top_Vis_Laser_Black = 1;
-        tacgt_ANPEQ_15_Top_Light_Tan = 1;
-        tacgt_ANPEQ_15_Top_Low_Light_Tan = 1;
-        tacgt_ANPEQ_15_Top_Laser_Tan = 1;
-        tacgt_ANPEQ_15_Top_Vis_Laser_Tan = 1;
-        tacgt_ANPEQ_15_Top_Light_OD = 1;
-        tacgt_ANPEQ_15_Top_Low_Light_OD = 1;
-        tacgt_ANPEQ_15_Top_Laser_OD = 1;
-        tacgt_ANPEQ_15_Top_Vis_Laser_OD = 1;
-        tacgt_LLM01_Laser_Black = 1;
-        tacgt_LLM01_Light_Black = 1;
-        tacgt_LLM01_Low_Light_Black = 1;
-        tacgt_LLM01_Vis_Laser_Black = 1;
-        tacgt_LLM_Laser_Black = 1;
-        tacgt_LLM_Light_Black = 1;
-        tacgt_LLM_Low_Light_Black = 1;
-        tacgt_LLM_Vis_Laser_Black = 1;
-        tacgt_Surefire_Flashlight_Black = 1;
-        tacgt_Surefire_Low_Flashlight_Black = 1;
-        tacgt_Surefire_Flashlight_OD = 1;
-        tacgt_Surefire_Low_Flashlight_OD = 1;
-        tacgt_Surefire_Flashlight_Tan = 1;
-        tacgt_UTG_Flashlight = 1;
-        tacgt_UTG_Low_Flashlight = 1;
-    };
-};
-
-class asdg_PistolUnderRail: asdg_SlotInfo {
-    class compatibleItems {
-        tacgt_CZ_M3X_Laser = 1;
-        tacgt_CZ_M3X_Light = 1;
-        tacgt_CZ_M3X_Vis_Laser = 1;
-        tacgt_MLPLS_Laser = 1;
-        tacgt_MLPLS_Vis_Laser = 1;
-    };
-};
-
+#include "joint_rails.hpp"
 #include "CfgWeapons.hpp"
 #include "CfgVehicles.hpp"
