+++ addons/flashlight/script_component.hpp
@@ -3,19 +3,17 @@
 #include "\x\tacgt\addons\main\script_mod.hpp"
 #include "\x\tacgt\addons\main\script_macros.hpp"
 
-#define MACRO_MULTI_1 \
+#define MACRO_PRIMARY \
     descriptionShort = "Laser/Light GT Edition"; \
     author = ECSTRING(main,Author); \
     scope = 2; \
-    scopeArsenal = 2;
+    scopeArsenal = 2
 
-#define MACRO_MULTI_2 \
-    descriptionShort = "Laser/Light GT Edition"; \
-    author = ECSTRING(main,Author); \
+#define MACRO_SECONDARY \
     scope = 1; \
-    scopeArsenal = 1;
+    scopeArsenal = 1
 
-#define MACRO_HI_FLASHLIGHT_SETTINGS \
+#define MACRO_FLASHLIGHT \
     color[] = {185, 190, 199}; \
     ambient[] = {6, 9, 9}; \
     intensity = 100; \
@@ -37,9 +35,9 @@
         hardLimitStart = 200; \
         hardLimitEnd = 300; \
     }; \
-    scale [] = {0};
+    scale [] = {0}
 
-#define MACRO_LOW_FLASHLIGHT_SETTINGS \
+#define MACRO_RED_FLASHLIGHT \
     color[] = {175, 50, 36}; \
     ambient[] = {9, 9, 9}; \
     intensity = 17; \
@@ -61,34 +59,21 @@
         hardLimitStart = 200; \
         hardLimitEnd = 300; \
     }; \
-    scale [] = {0};
+    scale [] = {0}
 
-#define MACRO_IR_POINTER_SETTINGS \
+#define MACRO_IR_POINTER \
     RMBhint = "Laser Pointer"; \
     irLaserPos = "laser pos"; \
     irLaserEnd = "laser dir"; \
-    irDistance = 5;
+    irDistance = 5
 
-#define MACRO_VIS_POINTER_SETTINGS \
-    color[] = {0, 0, 0}; \
-    ambient[] = {0, 0, 0}; \
-    intensity = 0; \
-    size = 0; \
-    innerAngle = 0; \
-    outerAngle = 0; \
-    coneFadeCoef = 0; \
-    position = "flash dir"; \
-    direction = "flash"; \
-    useFlare = 0; \
-    flareSize = 0; \
-    flareMaxDistance = 0; \
-    dayLight = 0; \
-    class Attenuation { \
-        start = 0; \
-        constant = 0; \
-        linear = 0; \
-        quadratic = 0; \
-        hardLimitStart = 0; \
-        hardLimitEnd = 0; \
-    }; \
-    scale [] = {0};
+#define MACRO_VISIBLE_POINTER \
+    irDistance = 5; \
+    irLaserEnd = "laser dir"; \
+    irLaserPos = "laser pos"; \
+    isIR = 0; \
+    irDotSize = 0.1; \
+    beamThickness = 0; \
+    beamMaxLength = 1000; \
+    beamColor[] = {0, 0, 0}; \
+    dotColor[] = {0, 10000, 0}
